# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Finance MVP API
  version: 0.1.0
  summary: Canonical OpenAPI contract for the mounted closed-MVP runtime surface.
  license:
    name: Proprietary/Internal
    url: https://example.invalid/proprietary-internal-license
  description: |
    This document is the source of truth for the mounted MVP wire routes, DTO field names,
    enum values, error codes, pagination envelopes, and auth requirements.

    MVP boundaries are intentionally narrow: financial data is manual-entry only,
    `sourceType` accepts only `manual`, and transfers are same-scope only. Safe
    screenshot OCR capture may create pending drafts from sanitized structured
    candidates, but raw image bytes and OCR text are rejected and confirmed
    transactions remain `sourceType=manual`. There is no real import, parser,
    file storage, bank API, automated ingestion, broker credential, external
    credential, card, IBAN, raw statement, or device-token endpoint.

    Privacy invariants:
    - personal data is owner-only;
    - shared data is visible only to active members of the same Household;
    - reports support `personal`, `shared_family_report`, and `combined_viewer_overview`;
    - every list, report, search, autocomplete, cache, count, and aggregation
      filters visible rows before any aggregation or pagination;
    - responses never expose hidden counts, filtered-out counts, hidden facets,
      or diagnostics that confirm inaccessible resources.
servers:
...
security:
  - pwaCookieAuth: []
    csrfHeader: []
  - androidBearerAuth: []
paths:
...
  /asset-categories:
    get:
      tags: [Asset Categories]
      operationId: listAssetCategories
      summary: List/search visible asset categories.
      description: Personal asset categories are owner-only; household asset categories require active membership. This is the source of truth for manual asset groups and investment flags.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: scopeType
          in: query
          schema:
            $ref: '#/components/schemas/AssetCategoryScope'
        - $ref: '#/components/parameters/HouseholdIdFilter'
        - $ref: '#/components/parameters/AssetCategoryRecordStatusFilter'
        - name: isInvestment
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: Visible asset categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryPageEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Asset Categories]
      operationId: createAssetCategory
      summary: Create personal or household asset category.
      description: Defaults `manualAmount` to 0 and `isInvestment` to false.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCategoryCreateRequest'
      responses:
        '201':
          description: Created asset category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /asset-categories/investment-migrations:
    post:
      tags: [Asset Categories]
      operationId: createInvestmentMigration
      summary: Atomically create an investment category and link accounts.
      description: |
        Online immediate command for converting active same-scope, same-currency,
        unlinked accounts into one investment asset category. The command is
        atomic within the backend request. Offline retry/idempotency is provided
        through `/sync/push` with `entityType=investment_migrations` and
        `operation=create`; this REST endpoint does not store clientMutationId.
        `color` is accepted for client command parity but is not persisted by the
        current asset category storage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestmentMigrationCreateRequest'
      responses:
        '201':
          description: Created investment category and linked accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestmentMigrationEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /asset-categories/{assetCategoryId}:
    get:
      tags: [Asset Categories]
      operationId: getAssetCategory
      summary: Get visible asset category detail.
      parameters:
        - $ref: '#/components/parameters/AssetCategoryId'
      responses:
        '200':
          description: Asset category detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Asset Categories]
      operationId: updateAssetCategory
      summary: Update asset category mutable fields.
      parameters:
        - $ref: '#/components/parameters/AssetCategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCategoryUpdateRequest'
      responses:
        '200':
          description: Updated asset category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Asset Categories]
      operationId: deleteAssetCategory
      summary: Soft delete asset category.
      parameters:
        - $ref: '#/components/parameters/AssetCategoryId'
      responses:
        '204':
          description: Asset category soft-deleted.
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /asset-categories/{assetCategoryId}/archive:
    post:
      tags: [Asset Categories]
      operationId: archiveAssetCategory
      summary: Archive asset category.
      parameters:
        - $ref: '#/components/parameters/AssetCategoryId'
      responses:
        '200':
          description: Archived asset category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /asset-categories/{assetCategoryId}/restore:
    post:
      tags: [Asset Categories]
      operationId: restoreAssetCategory
      summary: Restore archived asset category.
      parameters:
        - $ref: '#/components/parameters/AssetCategoryId'
      responses:
        '200':
          description: Restored asset category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetCategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List/search visible categories.
      description: Personal categories are owner-only; household categories require active membership. Usage counts for hidden personal activity are never exposed.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/CategoryScopeFilter'
        - $ref: '#/components/parameters/CategoryTypeFilter'
        - $ref: '#/components/parameters/HouseholdIdFilter'
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Visible categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryPageEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Categories]
      operationId: createCategory
      summary: Create personal or household category.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Created category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /categories/{categoryId}:
    get:
      tags: [Categories]
      operationId: getCategory
      summary: Get visible category detail.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Categories]
      operationId: updateCategory
      summary: Update category mutable fields.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Soft delete category if history remains coherent.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Category soft-deleted.
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /categories/{categoryId}/archive:
    post:
      tags: [Categories]
      operationId: archiveCategory
      summary: Archive category.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Archived category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /categories/{categoryId}/restore:
    post:
      tags: [Categories]
      operationId: restoreCategory
      summary: Restore category.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Restored category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    pwaCookieAuth:
      type: apiKey
      in: cookie
      name: __Host-finance_session
      description: HttpOnly Secure SameSite cookie used by the PWA.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for PWA cookie-authenticated state-changing requests; bound to the authenticated session and rotated on login/logout/reset.
    androidBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token for Android; refresh token storage and rotation are platform secure-storage concerns.
  responses:
    Error:
      description: Canonical error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  parameters:
...
    AssetCategoryId:
      name: assetCategoryId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        minLength: 1
      description: Endpoint-specific allowlisted sort value.
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Search runs only after visible-scope filtering and is not logged as raw payload.
...
    HouseholdIdFilter:
      name: householdId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
      description: Optional scope narrowing only; inaccessible ids use neutral behavior.
...
    CategoryScopeFilter:
      name: scope
      in: query
      required: false
      schema: { $ref: '#/components/schemas/CategoryScope' }
    CategoryTypeFilter:
      name: type
      in: query
      required: false
      schema: { $ref: '#/components/schemas/CategoryType' }
...
    StatusFilter:
      name: status
      in: query
      required: false
      schema: { $ref: '#/components/schemas/RecordStatus' }
    AssetCategoryRecordStatusFilter:
      name: recordStatus
      in: query
      required: false
      schema: { $ref: '#/components/schemas/RecordStatus' }
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9_:-]+$'
      examples: [usr_123]
    DateTime:
      type: string
      format: date-time
      examples: ['2026-05-17T10:00:00Z']
...
    DecimalString:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,6})?$'
      examples: ['1000.00']
      description: Decimal-safe money string; never floating point.
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      examples: [RUB]
    OwnershipType:
      type: string
      enum: [personal, shared]
    AccountType:
      type: string
      enum: [cash, bank, card, deposit, brokerage, metal, other]
    AssetCategoryType:
      type: string
      enum: [cash, bank, card, deposit, brokerage, metal, other]
    AssetCategoryScope:
      type: string
      enum: [personal, household]
    CategoryScope:
      type: string
      enum: [personal, household]
      description: Category scope uses `household`, not `shared`.
    CategoryType:
      type: string
      enum: [income, expense]
...
    RecordStatus:
      type: string
      enum: [active, archived, deleted]
...
    ErrorDto:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe generic user-facing message.
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDto'
    PageInfo:
      type: object
      additionalProperties: false
      required: [limit, hasMore]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        nextCursor:
          type: [string, 'null']
          description: Opaque cursor scoped to visible filtered rows only.
        hasMore:
          type: boolean
      description: Pagination metadata never includes totalCount or hidden counts.
...
    AccountDto:
      type: object
      additionalProperties: false
      required: [id, name, accountType, ownershipType, currency, initialBalance, currentBalance, isPaymentAccount, status, createdByUserId, createdAt, updatedAt, version]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          maxLength: 120
        accountType:
          $ref: '#/components/schemas/AccountType'
        ownershipType:
          $ref: '#/components/schemas/OwnershipType'
        ownerUserId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
          description: Required for personal accounts; null for shared accounts.
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
          description: Required for shared accounts; null for personal accounts.
        assetCategoryId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
          description: Optional link to an active same-scope same-currency asset category.
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        initialBalance:
          $ref: '#/components/schemas/DecimalString'
        currentBalance:
          $ref: '#/components/schemas/DecimalString'
        isPaymentAccount:
          type: boolean
          description: True when the account can be selected as an expense payment source, including capture-confirmed expenses. Income transactions are not blocked by this flag.
        status:
          $ref: '#/components/schemas/RecordStatus'
        createdByUserId:
          $ref: '#/components/schemas/ResourceId'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        archivedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        deletedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        version:
          type: integer
          minimum: 1
...
    AssetCategoryDto:
      type: object
      additionalProperties: false
      required: [id, name, scopeType, currency, assetType, manualAmount, isInvestment, recordStatus, createdByUserId, createdAt, updatedAt, version]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          maxLength: 120
        scopeType:
          $ref: '#/components/schemas/AssetCategoryScope'
        ownerUserId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        assetType:
          $ref: '#/components/schemas/AssetCategoryType'
        iconKey:
          type: [string, 'null']
          maxLength: 80
        manualAmount:
          $ref: '#/components/schemas/DecimalString'
        isInvestment:
          type: boolean
        recordStatus:
          $ref: '#/components/schemas/RecordStatus'
        createdByUserId:
          $ref: '#/components/schemas/ResourceId'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        archivedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        deletedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        version:
          type: integer
          minimum: 1
    AssetCategoryCreateRequest:
      type: object
      additionalProperties: false
      required: [name, scopeType, currency]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        scopeType:
          $ref: '#/components/schemas/AssetCategoryScope'
        householdId:
          $ref: '#/components/schemas/ResourceId'
          description: Required for household scope; forbidden for personal.
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        assetType:
          $ref: '#/components/schemas/AssetCategoryType'
          default: other
        iconKey:
          type: [string, 'null']
          minLength: 1
          maxLength: 80
        manualAmount:
          $ref: '#/components/schemas/DecimalString'
          default: '0.0000'
        isInvestment:
          type: boolean
          default: false
    AssetCategoryUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        manualAmount:
          $ref: '#/components/schemas/DecimalString'
        assetType:
          $ref: '#/components/schemas/AssetCategoryType'
        iconKey:
          type: [string, 'null']
          minLength: 1
          maxLength: 80
        isInvestment:
          type: boolean
        recordStatus:
          $ref: '#/components/schemas/RecordStatus'
        version:
          type: integer
          minimum: 1
    InvestmentMigrationCreateRequest:
      type: object
      additionalProperties: false
      required: [assetCategoryId, name, assetType, currency, accountIds, accountVersions]
      properties:
        assetCategoryId:
          $ref: '#/components/schemas/ResourceId'
          description: Client-provided ID for the asset category to create. For sync this must match mutation.entityId.
        name:
          type: string
          minLength: 1
          maxLength: 120
        icon:
          type: [string, 'null']
          minLength: 1
          maxLength: 80
          description: Stored as asset category iconKey. Backend also accepts iconKey.
        color:
          type: [string, 'null']
          minLength: 1
          maxLength: 32
          description: Accepted for client command parity; not persisted by current asset category storage.
        assetType:
          $ref: '#/components/schemas/AssetCategoryType'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        scope:
          $ref: '#/components/schemas/AssetCategoryScope'
          description: Optional. If omitted, derived from the validated accounts; backend also accepts scopeType.
        householdId:
          $ref: '#/components/schemas/ResourceId'
          description: Required for household scope unless derivable from shared accounts; forbidden for personal.
        accountIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/ResourceId'
        accountVersions:
          type: object
          additionalProperties:
            type: integer
            minimum: 1
          description: Map of accountId to expected optimistic version. Keys must exactly match accountIds.
      description: |
        Creates an investment asset category with manualAmount=0 and isInvestment=true,
        then links all requested accounts without changing balances. Rejects stale
        account versions, linked accounts, mixed scope, mixed currency, or account
        type versus assetType mismatches before writing.
...
    CategoryDto:
      type: object
      additionalProperties: false
      required: [id, name, type, scope, status, createdByUserId, createdAt, updatedAt, version]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          maxLength: 120
        type:
          $ref: '#/components/schemas/CategoryType'
        scope:
          $ref: '#/components/schemas/CategoryScope'
        ownerUserId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        iconKey:
          type: [string, 'null']
          maxLength: 80
        color:
          type: [string, 'null']
          pattern: '^#[0-9A-Fa-f]{6}$'
        status:
          $ref: '#/components/schemas/RecordStatus'
        createdByUserId:
          $ref: '#/components/schemas/ResourceId'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        archivedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        deletedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        version:
          type: integer
          minimum: 1
    CategoryCreateRequest:
      type: object
      additionalProperties: false
      required: [name, type, scope]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        type:
          $ref: '#/components/schemas/CategoryType'
        scope:
          $ref: '#/components/schemas/CategoryScope'
        householdId:
          $ref: '#/components/schemas/ResourceId'
          description: Required for household categories.
        iconKey:
          type: string
          maxLength: 80
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    CategoryUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        iconKey:
          type: string
          maxLength: 80
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        status:
          $ref: '#/components/schemas/RecordStatus'
        version:
          type: integer
          minimum: 1
      description: scope, ownerUserId, and householdId are immutable in MVP.
...
    AssetCategoryEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AssetCategoryDto'
    AssetCategoryPageEnvelope:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssetCategoryDto'
        page:
          $ref: '#/components/schemas/PageInfo'
    InvestmentMigrationDto:
      type: object
      additionalProperties: false
      required: [assetCategory, accounts]
      properties:
        assetCategory:
          $ref: '#/components/schemas/AssetCategoryDto'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountDto'
    InvestmentMigrationEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/InvestmentMigrationDto'
...
    CategoryEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/CategoryDto'
    CategoryPageEnvelope:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
        page:
          $ref: '#/components/schemas/PageInfo'
...